made the shape integer based (for the most part - full cleanup pass will come later)

+++ python/source/util.cpp
@@ -69,12 +69,12 @@ bind_yield_t<soil::shape_t<3>::arr_t>(module, "yield_shape_t_arr_3");
 //
 
 auto shape = py::class_<soil::shape>(module, "shape");
-shape.def(py::init<const std::vector<size_t>&>());
+shape.def(py::init<const std::vector<int>&>());
 
 shape.def("dims", &soil::shape::dims);
 shape.def("elem", &soil::shape::elem);
 shape.def("iter", &soil::shape::iter);
-shape.def("flat", [](const soil::shape& shape, const std::vector<size_t>& v){
+shape.def("flat", [](const soil::shape& shape, const std::vector<int>& v){
   if(v.size() == 1) return shape.flat<1>({v[0]});
   if(v.size() == 2) return shape.flat<2>({v[0], v[1]});
   if(v.size() == 3) return shape.flat<3>({v[0], v[1], v[2]});
@@ -103,7 +103,7 @@ shape.def("__repr__", [](const soil::shape& shape){
 
 auto array = py::class_<soil::array>(module, "array", py::buffer_protocol());
 array.def(py::init<const std::string, const soil::shape&>());
-array.def(py::init<>([](const std::string type, const std::vector<size_t>& v){
+array.def(py::init<>([](const std::string type, const std::vector<int>& v){
   auto shape = soil::shape(v);
   return soil::array(type, shape);
 }));
@@ -161,9 +161,9 @@ array.def("__setitem__", &soil::array::set<double>);
 
 array.def("__setitem__", [](soil::array& array, const py::tuple& tup, const py::object value){
   size_t index;
-  if(tup.size() == 1) index = array.shape().template flat<1>({tup[0].cast<size_t>()});
-  if(tup.size() == 2) index = array.shape().template flat<2>({tup[0].cast<size_t>(), tup[1].cast<size_t>()});
-  if(tup.size() == 3) index = array.shape().template flat<3>({tup[0].cast<size_t>(), tup[1].cast<size_t>(), tup[2].cast<size_t>()});
+  if(tup.size() == 1) index = array.shape().template flat<1>({tup[0].cast<int>()});
+  if(tup.size() == 2) index = array.shape().template flat<2>({tup[0].cast<int>(), tup[1].cast<int>()});
+  if(tup.size() == 3) index = array.shape().template flat<3>({tup[0].cast<int>(), tup[1].cast<int>(), tup[2].cast<int>()});
 
   if(array.type() == "int") array.set<int>(index, value.cast<int>());
   if(array.type() == "float") array.set<float>(index, value.cast<float>());
