+++ source/io/tiff.hpp
@@ -99,7 +99,7 @@ bool tiff::read(const char* filename){
   // Note: TIFF is Column Major (See Reading Function Below)
   //  Therefore,
   
-  auto shape = soil::shape({this->height(), this->width()});
+  auto shape = soil::shape({(int)this->height(), (int)this->width()});
 
   if(this->bits() == 32){
     this->_array = soil::array("float", shape);
